+++ components/search/index.d.ts
@@ -13,7 +13,7 @@ interface Props {
   placeholder?: string
   data?: DataItem
   onSearch?: (inputVal: string, item?: DataItem) => void
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
+  onChange?: (e: string) => void
   overlayClassName?: string
 }
 declare const Search: React.ComponentType<Props>